requirements.txt
certifi==2024.2.2
# dj-database-url==2.2.0
django==5.0.6
# psycopg2-binary==2.9.9
django-storages==1.14.3
# boto3==1.28.85
awsebcli==3.20.10
# gunicorn==22.0.0
# whitenoise==5.3.0


# Django==4.2.3
pip==23.3.2
wheel==0.40.0
gunicorn
whitenoise
psycopg2-binary
dj-database-url
chardet
# django-storages==1.14.2
boto3
pdfrw
pyPDF2
reportlab
xhtml2pdf